requirements.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
boto3==1.15.11
botocore==1.18.11
cffi==1.14.3
colorama==0.4.3
cryptography==3.1.1
Django==3.1.2
django-cors-headers==3.5.0
django-grappelli==2.14.2
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
djangorestframework-camel-case==1.2.0
Fabric3==1.14.post1
gunicorn==20.0.4
isort==5.5.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.2
Pillow==8.0.0
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
